+++ o365spray/core/handlers/enumerator/enumerate.py
@@ -62,6 +62,7 @@ def enumerate(args: argparse.Namespace, output_dir: str) -> object:
         writer=True,
         sleep=args.sleep,
         jitter=args.jitter,
+        proxy_url=args.proxy_url,
     )
 
     def enum_signal_handler(signal, frame):