+++ o365spray/core/handlers/sprayer/modules/base.py
@@ -43,6 +43,7 @@ class SprayerBase(BaseHandler):
         writer: bool = True,
         sleep: int = 0,
         jitter: int = 0,
+        proxy_url: str = None,
         *args,
         **kwargs,
     ):
@@ -69,6 +70,7 @@ class SprayerBase(BaseHandler):
             writer: toggle writing to output files
             sleep: throttle http requests
             jitter: randomize throttle
+            proxy_url: fireprox api url
 
         Raises:
             ValueError: if no output directory provided when output writing
@@ -91,6 +93,7 @@ class SprayerBase(BaseHandler):
         self.adfs_url = adfs_url
         self.sleep = sleep
         self.jitter = jitter
+        self.proxy_url = proxy_url
         self.executor = concurrent.futures.ThreadPoolExecutor(max_workers=workers)
 
         # Internal exit handler