+++ o365spray/core/handlers/validator/modules/getuserrealm.py
@@ -3,6 +3,10 @@
 import html
 import xml.etree.ElementTree as ET
 from typing import Tuple
+from o365spray.core.utils import (
+    Defaults,
+    Helper,
+)
 from o365spray.core.handlers.validator.modules.base import ValidatorBase
 
 
@@ -23,13 +27,24 @@ class ValidateModule_getuserrealm(ValidatorBase):
             (if the domain is a valid O365 domain, ADFS AuthURL)
         """
         (valid, adfs_url) = (False, None)  # Defaults
+        headers = Defaults.HTTP_HEADERS
 
-        url = f"https://login.microsoftonline.com/getuserrealm.srf?login=user@{domain}&xml=1"
+        # Handle FireProx API URL
+        if self.proxy_url:
+            proxy_url = self.proxy_url.rstrip("/")
+            url = f"{proxy_url}/getuserrealm.srf?login=user@{domain}&xml=1"
+
+            # Update headers
+            headers = Helper.fireprox_headers(headers)
+
+        else:
+            url = f"https://login.microsoftonline.com/getuserrealm.srf?login=user@{domain}&xml=1"
 
         # Send request
         rsp = self._send_request(
             "get",
             url,
+            headers=headers,
             proxies=self.proxies,
             timeout=self.timeout,
             sleep=self.sleep,