+++ o365spray/core/handlers/validator/validate.py
@@ -37,6 +37,7 @@ def validate(args: argparse.Namespace) -> argparse.Namespace:
         proxy=args.proxy,
         sleep=args.sleep,
         jitter=args.jitter,
+        proxy_url=args.proxy_url,
     )
     (valid, adfs) = v.validate(args.domain)